chore(components): rename authorize-to-twitch-button
